refactor(core,axum,proxy): one constant-time comparison, not two
`gglib-axum` and `gglib-proxy` each carried a private `constant_time_eq` with a
byte-identical body, and each guards bearer tokens with it. This is the
authentication path: a hardening change to one copy would not have been reported
against the other by any lint, because neither is dead and neither is public.

`normalize_host`, which both guards also call, already lives in
`gglib_core::access`, so the comparison joins it there beside `generate_api_key`
and `ApiKeySource`. Adding to the existing `mod.rs` means no module-table churn.

The tests move with it, since a unit test for a core function has no business
living in a surface crate. The proxy's block was the larger of the two — it adds
an empty-versus-nonempty case to the four assertions axum had, plus a separate
prefix test guarding the failure mode a naive `starts_with` would introduce — so
that is the one kept, and axum's are dropped rather than merged.
`gglib-proxy`'s `access` test module is left empty by the move and goes with it;
the guards' own behaviour is covered by `integration_auth` and `daemon_access`.

The guards themselves are deliberately not merged. Axum answers with
`Json(json!({… "type": "HOST_NOT_ALLOWED"}))` while the proxy answers with
`Json(ErrorResponse::with_code(…, "host_not_allowed"))`, over different state
types and different error envelopes. Only the comparison was literally
duplicated, and only it moves.

Going `pub` and cross-crate can cost an inlining opportunity but cannot introduce
an early exit, so no timing property changes. `#[must_use]` only adds a lint, and
both call sites consume the bool in an `if`.
